♻️refactor [#140] page파일아래 경로 수정, svg파일 이동 등 dashboard 경로 수정
