+++ src/components/icon/plus-circle-svg.tsx
