+++ src/components/icon/setting-svg.tsx
