+++ src/features/dashboard/edit/components/edit-sidebar.module.css
@@ -1,0 +1,110 @@
+.page {
+  display: flex;
+  flex-direction: column;
+  height: 100vh;
+  background-color: var(--color-background);
+}
+
+.page .topBar {
+  width: 100%;
+  background-color: var(--color-background);
+  border-bottom: 2px solid var(--color-stroke);
+  height: 72px;
+}
+
+.page .main {
+  display: flex;
+  height: 100%;
+}
+
+.page .main .sidebar {
+  width: 540px;
+  background-color: var(--color-black900);
+  position: fixed;
+  left: 0;
+  top: 0;
+  height: 100vh;
+  padding-top: 90px;
+  padding-right: 24px;
+  z-index: 100;
+}
+
+.page .main .contents {
+  flex: 1;
+  margin-left: 540px;
+  display: flex;
+  direction: row;
+}
+
+.page .main .sidebar .sidebarMenu {
+  width: 276px;
+  height: 175px;
+  display: flex;
+  flex-direction: column;
+  justify-self: flex-end;
+}
+
+.page .main .sidebar .sidebarMenu button {
+  background-color: transparent;
+  border: none;
+  color: #ffffff;
+  height: 58px;
+  text-align: left;
+  padding: 18.5px 14px 18.5px 14px;
+  border-radius: 12px;
+  cursor: pointer;
+  display: flex;
+  align-items: center;
+}
+
+.page .main .sidebar .sidebarMenu .betweenLine {
+  background-color: var(--color-modalBackground);
+  height: 1px;
+  width: 260px;
+}
+
+.page .main .sidebar .sidebarMenu button:last-child {
+  border-bottom: none;
+  display: flex;
+  justify-content: space-between;
+}
+
+.page .main .sidebar .sidebarMenu button:hover {
+  background-color: var(--color-black800);
+}
+
+.page .main .sidebar .sidebarMenu button:active {
+  background-color: var(--color-black700);
+}
+
+.page .main .sidebar .sidebarMenu button.activeButton {
+  background-color: #2c2b30;
+}
+
+.page .main .sidebar .sidebarMenu .trashButton {
+  color: var(--color-rose);
+}
+
+.page .main .contents .backButton {
+  padding-top: 30px;
+  display: flex;
+  flex-direction: column;
+  gap: 6px;
+  align-items: center;
+  cursor: pointer;
+}
+
+.page .main .contents .backButton span {
+  color: var(--color-gray300);
+  cursor: pointer;
+}
+
+.page .main .contents .backButton .xIcon {
+  width: 40px;
+  height: 40px;
+  color: transparent;
+}
+
+.page .main .contents .backButton:hover .xIcon {
+  color: var(--color-black500);
+}
