+++ src/pages/test-components/index.tsx
@@ -521,7 +521,7 @@ function BadgeSample() {
 function ColorPaletteSample() {
   const sizes = Object.values(ColorFrameSize);
   const [selectedColors, setSelectedColors] = useState<string[]>(
-    sizes.map(() => "")
+    sizes.map(() => ""),
   );
 
   const handleSelect = (index: number, color: string) => {
